setting up directory structure
