+++ src/Makefile
@@ -1,0 +1,7 @@
+CC=gcc++
+CFLAGS=-Wall -Wextra -std=c++11
+
+default: CardForge.o
+
+CardForge.o: CardForge.h
+	$(CC) $(CFLAGS) -c $< -I$(RAYLIBDIR) -L$(RAYLIBDIR)
