+++ test/uno_test/resources/blue_0.png
