+++ test/uno_test/resources/blue_1.png
