+++ test/uno_test/resources/blue_2.png
