+++ test/uno_test/resources/blue_3.png
