+++ test/uno_test/resources/blue_4.png
