+++ test/uno_test/resources/blue_5.png
