+++ test/uno_test/resources/blue_6.png
