+++ test/uno_test/resources/blue_7.png
