+++ test/uno_test/resources/blue_8.png
