+++ test/uno_test/resources/blue_9.png
