+++ test/uno_test/resources/blue_picker.png
