+++ test/uno_test/resources/blue_skip.png
