+++ test/uno_test/resources/green_6.png
