+++ test/uno_test/resources/green_picker.png
