+++ test/uno_test/resources/green_skip.png
