+++ test/uno_test/resources/red_0.png
