+++ test/uno_test/resources/red_2.png
