+++ test/uno_test/resources/red_3.png
