+++ test/uno_test/resources/red_6.png
