+++ test/uno_test/resources/red_7.png
