+++ test/uno_test/resources/red_8.png
