+++ test/uno_test/resources/red_picker.png
