+++ test/uno_test/resources/red_skip.png
