+++ test/uno_test/resources/wild_color_changer.png
