+++ test/uno_test/resources/wild_pick_four.png
