+++ test/uno_test/resources/yellow_0.png
