+++ test/uno_test/resources/yellow_1.png
