+++ test/uno_test/resources/yellow_2.png
