+++ test/uno_test/resources/yellow_3.png
