+++ test/uno_test/resources/yellow_4.png
