+++ test/uno_test/resources/yellow_5.png
