+++ test/uno_test/resources/yellow_6.png
