+++ test/uno_test/resources/yellow_7.png
