+++ test/uno_test/resources/yellow_8.png
