+++ test/uno_test/resources/yellow_9.png
