+++ test/uno_test/resources/yellow_picker.png
