+++ test/uno_test/resources/yellow_skip.png
